First scaling-up refactoring : we archive 610i on MBP M2. I will now moove arrays from stack to heap to allow even WIDER arrays

+++ SMS-Simulator/Individual.hpp
@@ -28,8 +28,8 @@ class Individual {
 	public:
         float gamma;
         float delta;
-        int agentid;
-		Individual(SexualMarket& world, int agentid);
+        unsigned long int agentid;
+		Individual(SexualMarket& world, unsigned long int agentid);
         akml::Matrix<float, P_DIMENSION, 1> getP();
         bool takeAction();
         bool responseToAction(Individual* from, float new_weight);
